Decouple update match modal
