+++ components/matches/match-list.module.scss
@@ -80,19 +80,6 @@
   }
 }
 
-.updateMatch {
-  display: flex;
-  flex-direction: column;
-
-  .formControl {
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 10px;
-  }
-}
-
-
 @media (max-width: $medium-small-device) {
   .section {
     flex-direction: column;
