+++ components/matches/modals/update-match.module.scss
@@ -1,0 +1,11 @@
+.updateMatch {
+  display: flex;
+  flex-direction: column;
+
+  .formControl {
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 10px;
+  }
+}
